More shaders ported the new system
Former-commit-id: 046b0c02a9a36a56751579e891ba2a289e6513df

+++ Source/BansheeUtility/Source/BsTextureAtlasLayout.cpp
@@ -119,7 +119,7 @@ namespace bs
 	{
 		for (size_t i = 0; i < elements.size(); i++)
 		{
-			elements[i].output.idx = i; // Preserve original index before sorting
+			elements[i].output.idx = (UINT32)i; // Preserve original index before sorting
 			elements[i].output.page = -1;
 		}
